+++ src/biforce/graphs/MatrixBipartiteSubgraph2.java
@@ -118,7 +118,7 @@ public class MatrixBipartiteSubgraph2 extends Subgraph2
      * @return 
      */
     @Override
-    public double edgeWeight(Vertex2 vtx1, Vertex2 vtx2) {
+    public float edgeWeight(Vertex2 vtx1, Vertex2 vtx2) {
         return supergraph.edgeWeight(vtx1, vtx2);
     }
     
@@ -184,7 +184,7 @@ public class MatrixBipartiteSubgraph2 extends Subgraph2
      * @param EdgeWeight 
      */
     @Override
-    public void setEdgeWeight(Vertex2 vtx1, Vertex2 vtx2, double edgeWeight) {
+    public void setEdgeWeight(Vertex2 vtx1, Vertex2 vtx2, float edgeWeight) {
         supergraph.setEdgeWeight(vtx1, vtx2, edgeWeight);
     }
