+++ src/biforce/graphs/MatrixHierGeneralSubgraph.java
@@ -116,7 +116,7 @@ public class MatrixHierGeneralSubgraph extends Subgraph2 implements
      * @return 
      */
     @Override
-    public double edgeWeight(Vertex2 vtx1, Vertex2 vtx2) {
+    public float edgeWeight(Vertex2 vtx1, Vertex2 vtx2) {
         return supergraph.edgeWeight(vtx1,vtx2);
     }
 
@@ -164,7 +164,7 @@ public class MatrixHierGeneralSubgraph extends Subgraph2 implements
     }
 
     @Override
-    public void setEdgeWeight(Vertex2 vtx1, Vertex2 vtx2, double edgeWeight) {
+    public void setEdgeWeight(Vertex2 vtx1, Vertex2 vtx2, float edgeWeight) {
         supergraph.setEdgeWeight(vtx1, vtx2, edgeWeight);
     }
